Health check endpoint and unit test for it
