+++ backend/src/api/health.rs
@@ -1,0 +1,6 @@
+use actix_web::{get, HttpResponse};
+
+#[get("/health_check")]
+pub async fn health_check() -> HttpResponse {
+    HttpResponse::Ok().finish()
+}
