+++ backend/src/api/mod.rs
@@ -1,2 +1,3 @@
 pub mod game;
 pub mod user;
+pub mod health;
